+++ presentation/src/main/kotlin/com/sooum/where_android/view/main/MainScreenView.kt
@@ -280,13 +280,13 @@ private fun NavHostController.navigateMeetDetail(meetDetail: MeetDetail) {
     navigateMeetDetailById(meetDetail.id)
 }
 
-private fun NavHostController.navigateMeetDetailById(id:Long) {
+private fun NavHostController.navigateMeetDetailById(id:Int) {
     //TODO : 환님 여기서 연결하면 됩니다.
     //
     context.startActivity(
         Intent(context, MyMeetActivity::class.java).apply {
             putExtras(Bundle().apply {
-                putLong(MyMeetActivity.MEET_ID, id)
+                putInt(MyMeetActivity.MEET_ID, id)
             })
         }
     )
